// packages/Files/src/types.ts
export interface Types {
  '7z': 'application/x-7z-compressed'
  aac: 'audio/aac'
  aif: 'audio/x-aiff'
  aifc: 'audio/x-aiff'
  aiff: 'audio/x-aiff'
  au: 'audio/basic'
  avi: 'video/x-msvideo'
  bmp: 'image/x-ms-bmp'
  bz: 'application/x-bzip'
  bz2: 'application/x-bzip2'
  csv: 'text/csv'
  doc: 'application/msword'
  docx: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document'
  gif: 'image/gif'
  gz: 'application/gzip'
  ico: 'image/x-icon'
  jpe: 'image/jpeg'
  jpeg: 'image/jpeg'
  jpg: 'image/jpeg'
  m1v: 'video/mpeg'
  mid: 'audio/midi'
  midi: 'audio/x-midi'
  mov: 'video/quicktime'
  movie: 'video/x-sgi-movie'
  mp2: 'audio/mpeg'
  mp3: 'audio/mpeg'
  mp4: 'video/mp4'
  mpa: 'video/mpeg'
  mpe: 'video/mpeg'
  mpeg: 'video/mpeg'
  mpg: 'video/mpeg'
  pbm: 'image/x-portable-bitmap'
  pdf: 'application/pdf'
  pgm: 'image/x-portable-graymap'
  png: 'image/png'
  pnm: 'image/x-portable-anymap'
  pot: 'application/vnd.ms-powerpoint'
  ppa: 'application/vnd.ms-powerpoint'
  ppm: 'image/x-portable-pixmap'
  pps: 'application/vnd.ms-powerpoint'
  ppt: 'application/vnd.ms-powerpoint'
  pptx: 'application/vnd.openxmlformats-officedocument.presentationml.presentation'
  ps: 'application/postscript'
  pwz: 'application/vnd.ms-powerpoint'
  qt: 'video/quicktime'
  ra: 'audio/x-pn-realaudio'
  rar: 'application/vnd.rar'
  ras: 'image/x-cmu-raster'
  rgb: 'image/x-rgb'
  snd: 'audio/basic'
  svg: 'image/svg+xml'
  tif: 'image/tiff'
  tiff: 'image/tiff'
  wav: 'audio/x-wav'
  weba: 'audio/webm'
  webm: 'video/webm'
  webp: 'image/webp'
  wiz: 'application/msword'
  xbm: 'image/x-xbitmap'
  xls: 'application/vnd.ms-excel'
  xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet'
  xpm: 'image/x-xpixmap'
  xwd: 'image/x-xwindowdump'
  zip: 'application/zip'
}

// https://stackoverflow.com/questions/66500729/union-type-of-keyof-and-string-with-working-autocompletion
// eslint-disable-next-line @typescript-eslint/ban-types
export type Type = Types[keyof Types] | (string & {})

interface TypesWithString extends Types {
  [key: string]: string
}

export const types: TypesWithString = {
  '7z': 'application/x-7z-compressed',
  aac: 'audio/aac',
  aif: 'audio/x-aiff',
  aifc: 'audio/x-aiff',
  aiff: 'audio/x-aiff',
  au: 'audio/basic',
  avi: 'video/x-msvideo',
  bmp: 'image/x-ms-bmp',
  bz: 'application/x-bzip',
  bz2: 'application/x-bzip2',
  csv: 'text/csv',
  doc: 'application/msword',
  docx: 'application/vnd.openxmlformats-officedocument.wordprocessingml.document',
  gif: 'image/gif',
  gz: 'application/gzip',
  ico: 'image/x-icon',
  jpe: 'image/jpeg',
  jpeg: 'image/jpeg',
  jpg: 'image/jpeg',
  m1v: 'video/mpeg',
  mid: 'audio/midi',
  midi: 'audio/x-midi',
  mov: 'video/quicktime',
  movie: 'video/x-sgi-movie',
  mp2: 'audio/mpeg',
  mp3: 'audio/mpeg',
  mp4: 'video/mp4',
  mpa: 'video/mpeg',
  mpe: 'video/mpeg',
  mpeg: 'video/mpeg',
  mpg: 'video/mpeg',
  pbm: 'image/x-portable-bitmap',
  pdf: 'application/pdf',
  pgm: 'image/x-portable-graymap',
  png: 'image/png',
  pnm: 'image/x-portable-anymap',
  pot: 'application/vnd.ms-powerpoint',
  ppa: 'application/vnd.ms-powerpoint',
  ppm: 'image/x-portable-pixmap',
  pps: 'application/vnd.ms-powerpoint',
  ppt: 'application/vnd.ms-powerpoint',
  pptx: 'application/vnd.openxmlformats-officedocument.presentationml.presentation',
  ps: 'application/postscript',
  pwz: 'application/vnd.ms-powerpoint',
  qt: 'video/quicktime',
  ra: 'audio/x-pn-realaudio',
  rar: 'application/vnd.rar',
  ras: 'image/x-cmu-raster',
  rgb: 'image/x-rgb',
  snd: 'audio/basic',
  svg: 'image/svg+xml',
  tif: 'image/tiff',
  tiff: 'image/tiff',
  wav: 'audio/x-wav',
  weba: 'audio/webm',
  webm: 'video/webm',
  webp: 'image/webp',
  wiz: 'application/msword',
  xbm: 'image/x-xbitmap',
  xls: 'application/vnd.ms-excel',
  xlsx: 'application/vnd.openxmlformats-officedocument.spreadsheetml.sheet',
  xpm: 'image/x-xpixmap',
  xwd: 'image/x-xwindowdump',
  zip: 'application/zip',
}